requirements.txt
requests
python-dotenv
colorama
prettytable